backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
PyJWT==2.8.0
httpx==0.25.2
requests==2.31.0
anthropic==0.7.0
stripe==7.0.0
pillow==10.1.0
python-multipart==0.0.6
aiofiles==23.2.1
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
sqlalchemy-utils==0.41.1
psycopg2-binary==2.9.9
mediapipe==0.10.9
opencv-python==4.8.1.78
numpy==1.24.3
biopython==1.81
pubmedpy
alembic==1.12.1
slowapi==0.1.9
redis>=4.5.2,<5.0.0
celery[redis]==5.3.4
sentry-sdk[fastapi]==1.40.0
python-json-logger==2.0.7
